refactor(dashboard): isolate navigation and farm canvas lifecycle
